# mediator/openapi/testdata/golden.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Orders",
    "summary": "The example service of spec 13",
    "description": "Exercises every generator feature.",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "API team",
      "url": "https://example.com",
      "email": "api@example.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.2.3"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/api/reports/": {
      "get": {
        "tags": [
          "reports"
        ],
        "operationId": "reports",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array",
                    "null"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (bad_request)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed (validation)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Internal error (internal)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "rule": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details. Every error response of the API has this shape.",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
...
}